// src/app/models/usuario.model.ts
export class UsuarioModel {
    _id?: string;
    strNombre: string;
    strPrimerApellido: string;
    strSegundoApellido?: string = '';
    strTelefono: string;
    strCorreo: string;
    strPassword?: string;
    strColonia: string;
    strCalle: string;
    nmbCodigoPostal: any;
    strCentroCrecer?: string;
    strImagen: string;
    ajsnRol?: [{

    }];
 }